Add UnitCode representation to Sensors.

+++ simulator/src/main/java/com/rles/simulator/enums/UnitCode.java
@@ -2,33 +2,33 @@ package com.rles.simulator.enums;
 
 // Represent unit codes as integers for byte frames
 public enum UnitCode {
-	NONE(0, ""),
-	PERCENT(1, "%"),
-	CELSIUS(2, "C"),
-	FAHRENHEIT(3, "F"),
-	KELVIN(4, "K"),
-	DEGREE(5, "Degree"),
-	METER(6, "m"),
-	METER_PER_SECOND(7, "m/s"),
-	AMP(8, "A"),
-	VOLT(9, "V"),
-	PASCAL(10, "Pa"),
-	CUBIC_METER_PER_SECOND(11, "m3/s"),
-	KILOGRAM_PER_SECOND(12, "kg/s"),
-	BOOLEAN(13, "bool");
+	NONE((byte)0, ""),
+	PERCENT((byte)1, "%"),
+	CELSIUS((byte)2, "C"),
+	FAHRENHEIT((byte)3, "F"),
+	KELVIN((byte)4, "K"),
+	DEGREE((byte)5, "Degree"),
+	METER((byte)6, "m"),
+	METER_PER_SECOND((byte)7, "m/s"),
+	AMP((byte)8, "A"),
+	VOLT((byte)9, "V"),
+	PASCAL((byte)10, "Pa"),
+	CUBIC_METER_PER_SECOND((byte)11, "m3/s"),
+	KILOGRAM_PER_SECOND((byte)12, "kg/s"),
+	BOOLEAN((byte)13, "bool");
 	
-	private final int code;
+	private final byte code;
 	private final String symbol;
 	
-	UnitCode(int code, String symbol) {
+	UnitCode(byte code, String symbol) {
 		this.code = code;
 		this.symbol = symbol;
 	}
 	
-	public int getCode() { return code; }
+	public byte getCode() { return code; }
 	public String getSymbol() { return symbol; }
 	
-	public static UnitCode getUnitCode(int code) {
+	public static UnitCode getUnitCode(byte code) {
 		for (UnitCode u : values()) {
 			if (u.code == code) return u;
 		}
